feat: added add funds endpoint, removed add card endpoint and revamped card model

+++ db/migrations/20230304211637_create-wallet-model.js
@@ -9,7 +9,7 @@ exports.up = function(knex) {
         table.string('account_number').notNullable().unique();
         table.string('account_name').notNullable().references('name').inTable('users');
         table.string('account_bank').notNullable();
-        table.integer('balance').notNullable();
+        table.decimal('balance').notNullable();
         table.boolean('is_active').defaultTo(false);
         table.timestamp('createdAt').notNullable().defaultTo(knex.fn.now());
         table.timestamp('updatedAt').notNullable().defaultTo(knex.fn.now());
