+++ models/wallet.model.js
@@ -6,7 +6,7 @@ const walletSchema = knex.schema.createTable('wallets', (table) => {
   table.string('account_number').notNullable().unique();
   table.string('account_name').notNullable().references('name').inTable('users');
   table.string('account_bank').notNullable();
-  table.integer('balance').notNullable();
+  table.decimal('balance').notNullable();
   table.boolean('is_active').defaultTo(false);
   table.timestamp('createdAt').notNullable().defaultTo(knex.fn.now());
   table.timestamp('updatedAt').notNullable().defaultTo(knex.fn.now());
